+++ src/MetaheuristicsPlatform/Algorithms/HyperHeuristics/FrrmabHyperHeuristic/FrrmabHyperHeuristicOptimizerReferences.cs
@@ -1,0 +1,14 @@
+using MetaheuristicsPlatform.Classification;
+
+namespace MetaheuristicsPlatform.Algorithms.HyperHeuristics.FrrmabHyperHeuristic;
+
+public static class FrrmabHyperHeuristicOptimizerReferences
+{
+    public static ScientificReference Primary { get; } =
+        new(
+            "Li, Fialho, Kwong & Zhang",
+            2014,
+            "Adaptive Operator Selection With Bandits for a Multiobjective Evolutionary Algorithm Based on Decomposition",
+            "IEEE Transactions on Evolutionary Computation 18(1), 114-130",
+            "10.1109/TEVC.2013.2239648");
+}